+++ enclave-manager/web/packages/app/src/emui/enclaves/EnclaveList.tsx
@@ -1,7 +1,6 @@
 import { Button, ButtonGroup, Flex } from "@chakra-ui/react";
 import { AppPageLayout, KurtosisAlert, PageTitle } from "kurtosis-ui-components";
 import { useEffect, useMemo, useState } from "react";
-import { useExperiments } from "../experiments/ExperimentsContext";
 import { BrowserRecommendator } from "./components/BrowserRecommendator";
 import { KurtosisUpgrader } from "./components/KurtosisUpgrader";
 import { EnclavesTable } from "./components/tables/EnclavesTable";
@@ -15,7 +14,6 @@ export const EnclaveList = () => {
 
   const [selectedEnclaves, setSelectedEnclaves] = useState<EnclaveFullInfo[]>([]);
 
-  const { experiments } = useExperiments();
   const enclavesKey = useMemo(
     () =>
       enclaves.isErr
@@ -48,7 +46,7 @@ export const EnclaveList = () => {
         </Flex>
       </Flex>
       <Flex direction="column" pt={"24px"} width={"100%"} gap={8}>
-        {experiments.enableCloudVersionUpgrade && <KurtosisUpgrader />}
+        <KurtosisUpgrader />
         <BrowserRecommendator />
         {enclaves.isOk && (
           <EnclavesTable